arm: Extract arm_undefined_exception C code to a separate file
